requirements.txt
pydantic==2.10.4
pydantic_core==2.27.2
pyFUME==0.2.25
PyGithub==1.58.1
PyGithub==1.58.1
openai==1.59.4
numpy==1.26.4
pandas==2.2.3
faiss-cpu==1.9.0.post1
huggingface-hub==0.27.1
Markdown==3.4.1
markdown2==2.5.2
tensorboard==2.10.1
sentence-transformers==3.3.1
asyncio==3.4.3
semgrep==1.112.0
tree-sitter==0.24.0
tree-sitter-languages==1.10.2
tree-sitter-language-pack==0.7.2
tenacity==9.0.0
Flask==3.1.0
GitPython==3.1.44
PyYAML==6.0.2
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.1.0
